+++ src/TemplatePreview/errors/PermissionDeniedToTemplate.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import {translate} from 'react-i18next';
-import {colors, Icon} from '@cimpress/react-components';
+import {colors} from '@cimpress/react-components';
+import {IconAlertTriangle} from '@cimpress-technology/react-streamline-icons';
 import {RobotCard} from '../../internal/RobotCard';
 import PropTypes from 'prop-types';
+import {getI18nInstance} from '../../i18n';
 
 class PermissionDeniedToTemplate extends React.Component {
     tt(key) {
@@ -15,7 +17,7 @@ class PermissionDeniedToTemplate extends React.Component {
             <RobotCard bsStyle={'danger'}>
                 <br/>
                 <h4>
-                    <Icon name={'report-problem-triangle-f'} size={'2x'} color={colors.danger.base}/>&nbsp;
+                    <IconAlertTriangle size={'2x'} style={{verticalAlign: 'middle'}} color={colors.danger.base}/>&nbsp;
                     {this.tt('errors:no_access_to_this_template_title')}
                 </h4>
                 <br/>
@@ -35,4 +37,4 @@ PermissionDeniedToTemplate.propTypes = {
     t: PropTypes.any,
 };
 
-export default translate(['translations', 'errors'])(PermissionDeniedToTemplate);
+export default translate(['translations', 'errors'], {i18n: getI18nInstance()})(PermissionDeniedToTemplate);