+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2019-11-14
+### Changed
+- Renamed package to @cimpress-technology/react-cimpress-templates
+- Replaced icons with equivalents from @cimpress/technology-react-streamline-icons
+
 ## [0.4.1] - 2019-10-21
 ### Changed
 - Handle case when clicking on a hyperlink with a span within it.